+++ libpp/profile_container.cpp
@@ -74,7 +74,7 @@ void profile_container::add(profile_t const & profile,
 
 	for (symbol_index_t i = 0; i < abfd.syms.size(); ++i) {
 
-		u32 start, end;
+		unsigned long start, end;
 		symbol_entry symb_entry;
 
 		abfd.get_symbol_range(i, start, end);
